# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Script Master Labs \u2014 SqueezeOS, 402Proof & Ghost Layer API",
    "version": "2.0.0",
    "description": "Three-system platform: SqueezeOS (institutional AI trading intelligence), 402Proof (x402/HTTP 402 autonomous agent payment compliance firewall on XRPL), and Ghost Layer (dual-chain transparent toll gateway bridging XRPL and Base/EVM). All premium SqueezeOS endpoints follow the x402 standard \u2014 agents pay RLUSD on XRPL, no API keys required. Ghost Layer routes XRPL and Base chain USDC payments with full fee disclosure. Agent SDK: https://four02proof.onrender.com/agent/",
    "contact": {
      "name": "Script Master Labs",
      "url": "https://scriptmasterlabs.com/terminal"
    },
    "x-llm-summary": "Two-system platform: SqueezeOS delivers real-time market intelligence (squeeze scanner, options flow, IWM 0DTE, AI council verdicts). 402Proof handles x402-compliant RLUSD micropayment settlement on XRPL so autonomous agents can pay per call with no API keys. Loyalty tiers reward cumulative spend. MCP middleware available via npm @relayos/mcp-paywall."
  },
  "servers": [
    {
      "url": "https://four02proof.onrender.com",
      "description": "402Proof \u2014 Payment Settlement Server (x402 compliant, Agent Passport, compliance receipts)"
    },
    {
      "url": "https://squeezeos-terminal.vercel.app",
      "description": "SqueezeOS \u2014 Market Intelligence API (Railway, x402 gated)"
    },
    {
...
    }
  ],
  "paths": {
...
    "/api/futures/settle/{future_id}": {
      "post": {
        "summary": "Force-settle a future against latest verdict",
        "operationId": "futuresSettle",
        "tags": ["Futures"],
        "parameters": [
          {"name": "future_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "Settlement proof"},
          "400": {"description": "No verdict found or future not ACTIVE"}
        }
      }
    },
...
    "/api/settlement/create": {
      "post": {
        "summary": "Create a conditional settlement contract (free)",
        "description": "Lock intent to pay RLUSD to another agent when a condition is met. Conditions: bias_match, confidence_above, price_above, price_below, time_elapsed. Zero custody — 1% platform fee logged on settlement.",
        "operationId": "settlementCreate",
        "tags": ["Settlement"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["creator_wallet", "recipient_wallet", "amount_rlusd", "condition_type", "symbol"],
                "properties": {
                  "creator_wallet":   {"type": "string"},
                  "recipient_wallet": {"type": "string"},
                  "amount_rlusd":     {"type": "number", "minimum": 0.01, "maximum": 1000},
                  "condition_type":   {"type": "string", "enum": ["bias_match","confidence_above","price_above","price_below","time_elapsed"]},
                  "symbol":           {"type": "string"},
                  "condition_value":  {"type": "string", "description": "Threshold — e.g. 'BULLISH', '75', '220.50'"},
                  "description":      {"type": "string"},
                  "ttl_hours":        {"type": "integer", "default": 24}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Contract created"},
          "400": {"description": "Validation error"}
        }
      }
    },
    "/api/settlement": {
      "get": {
        "summary": "Browse open settlement contracts (free)",
        "operationId": "settlementList",
        "tags": ["Settlement"],
        "parameters": [
          {"name": "symbol", "in": "query", "schema": {"type": "string"}},
          {"name": "status", "in": "query", "schema": {"type": "string", "enum": ["OPEN","TRIGGERED","SETTLED","EXPIRED","CANCELLED"]}},
          {"name": "limit",  "in": "query", "schema": {"type": "integer", "default": 50}}
        ],
        "responses": {
          "200": {"description": "List of contracts"}
        }
      }
    },
    "/api/settlement/{contract_id}": {
      "get": {
        "summary": "Get settlement contract status",
        "operationId": "settlementGet",
        "tags": ["Settlement"],
        "parameters": [
          {"name": "contract_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "Contract detail"},
          "404": {"description": "Not found"}
        }
      }
    },
    "/api/settlement/trigger/{contract_id}": {
      "post": {
        "summary": "Check condition and settle if met",
        "description": "Anyone may call. Checks live data against contract condition. Emits SETTLEMENT_COMPLETE SSE event on success.",
        "operationId": "settlementTrigger",
        "tags": ["Settlement"],
        "parameters": [
          {"name": "contract_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "Condition met — settlement proof returned"},
          "202": {"description": "Condition not yet met"},
          "400": {"description": "Invalid state"}
        }
      }
    },
    "/api/settlement/cancel/{contract_id}": {
      "post": {
        "summary": "Cancel a contract (creator only, OPEN status)",
        "operationId": "settlementCancel",
        "tags": ["Settlement"],
        "parameters": [
          {"name": "contract_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["creator_wallet"],
                "properties": {
                  "creator_wallet": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Contract cancelled"},
          "403": {"description": "Not creator"},
          "400": {"description": "Not cancellable"}
        }
      }
    },
    "/api/settlement/wallet/{wallet}": {
      "get": {
        "summary": "All settlement contracts for a wallet",
        "operationId": "settlementWallet",
        "tags": ["Settlement"],
        "parameters": [
          {"name": "wallet", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Wallet contracts"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "ADMIN_TOKEN from Render environment. Generate with: openssl rand -hex 32. Required for /v1/admin/* endpoints."
      },
      "merchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Merchant API key obtained via POST /v1/merchant/register. Used for managing registered endpoints."
      },
      "x402Payment": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Payment-Token",
        "description": "JWT access token obtained via the 402Proof payment flow. Agents: hit the endpoint, handle the 402, pay RLUSD on XRPL, call /v1/verify, then retry with this header. Token TTL: 1 hour. Python SDK: pip install requests xrpl-py + use Proof402Client. NPM middleware: @relayos/mcp-paywall."
      }
    },
...
  },
...
}